+++ spec/platformApi/PlatformApiRestClient.spec.ts
@@ -1,7 +1,6 @@
 import chai from 'chai';
 import sinon from 'sinon';
-import { Logger } from '../../src';
-import { PlatformApiRestClient } from '../../src/platformApi/PlatformApiRestClient';
+import { NoAuthRestClient, Logger, PlatformApiRestClient } from '../../src';
 import * as flow from '../helpers/flow.json';
 
 const { expect } = chai;
@@ -73,5 +72,21 @@ describe('PlatformApiRestClient', () => {
     } catch (e) {
       expect(e.message).to.be.contains('Test case does not expect success response');
     }
+  });
+
+  it('Should succeed make request with user-agent Headers', async () => {
+    successBody = flow;
+    stub = sinon.stub(NoAuthRestClient.prototype, 'makeRequest');
+    stub.returns(successBody);
+    const requestBody = { url: 'https://example.com', method: 'GET', headers: { 'Content-Type': 'application/json' } };
+    const platformClient = new PlatformApiRestClient(emitter, cfg, 'UnitTest/1.0', 'messageId');
+    const result = await platformClient.makeRequest(requestBody);
+    expect(result).to.be.deep.equal(successBody);
+    expect(stub.called).to.be.equal(true);
+    expect(stub.firstCall.args[0].headers).to.be.deep.equal({
+      'Content-Type': 'application/json',
+      'User-Agent': 'UnitTest/1.0 request/2.88.2',
+      'x-request-id': 'f:undefined;s:undefined;m:messageId'
+    });
   });
 });